Move FlexConfig to sandbox location
