+++ src/ext/java/org/opentripplanner/ext/flex/FlexRouter.java
@@ -28,7 +28,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.opentripplanner.standalone.config.FlexConfig;
 
 public class FlexRouter {
