+++ src/ext/java/org/opentripplanner/ext/flex/template/FlexAccessEgressTemplate.java
@@ -20,7 +20,7 @@ import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
 import org.opentripplanner.routing.vertextype.TransitStopVertex;
-import org.opentripplanner.standalone.config.FlexConfig;
+import org.opentripplanner.ext.flex.FlexConfig;
 
 public abstract class FlexAccessEgressTemplate {
   protected final NearbyStop accessEgress;
